+++ utils/admin.cc
@@ -401,8 +401,6 @@ int main_impl(int argc, char** argv)
 
   std::shared_ptr<wampcc::wamp_session> ws;
 
-  auto ping_interval = wampcc::protocol::options::default_ping_interval;
-
   wampcc::wamp_session::options session_opts;
   session_opts.max_pending_open = uopts.timeout;
 
@@ -410,7 +408,6 @@ int main_impl(int argc, char** argv)
     case user_options::transport::websocket: {
       wampcc::websocket_protocol::options proto_opts;
       proto_opts.serialisers = uopts.serialisers;
-      proto_opts.ping_interval = ping_interval;
       ws = wampcc::wamp_session::create<wampcc::websocket_protocol>(
         g_kernel.get(),
         std::move(sock),
@@ -422,7 +419,6 @@ int main_impl(int argc, char** argv)
     case user_options::transport::rawsocket: {
       wampcc::rawsocket_protocol::options proto_opts;
       proto_opts.serialisers = uopts.serialisers;
-      proto_opts.ping_interval = ping_interval;
       ws = wampcc::wamp_session::create<wampcc::rawsocket_protocol>(
         g_kernel.get(),
         std::move(sock),